Move AnyCounterparty context to new hermes-any-counterparty crate
