+++ crates/any/any-counterparty/src/types/client_state.rs
@@ -1,5 +1,4 @@
 use hermes_cosmos_chain_components::types::tendermint::TendermintClientState;
-use ibc_relayer::client_state::AnyClientState as RelayerAnyClientState;
 use ibc_relayer_types::core::ics24_host::identifier::ChainId;
 use serde::Serialize;
 
@@ -8,23 +7,12 @@ use serde::Serialize;
 pub enum AnyClientState {
     Tendermint(TendermintClientState),
     // WasmTendermint(WasmTendermintClientState),
-    // Sovereign(SovereignClientState),
 }
 
 impl AnyClientState {
     pub fn chain_id(&self) -> &ChainId {
         match self {
             AnyClientState::Tendermint(client_state) => &client_state.chain_id,
-        }
-    }
-}
-
-impl From<RelayerAnyClientState> for AnyClientState {
-    fn from(client_state: RelayerAnyClientState) -> Self {
-        match client_state {
-            RelayerAnyClientState::Tendermint(client_state) => {
-                AnyClientState::Tendermint(client_state)
-            }
         }
     }
 }
