+++ crates/cli/cli-components/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod any_client;
 pub mod impls;
 pub mod traits;
