+++ crates/cosmos/cosmos-relayer/src/contexts/chain.rs
@@ -5,9 +5,9 @@ use core::ops::Deref;
 use cgp_core::error::{ErrorRaiserComponent, ErrorTypeComponent};
 use cgp_core::prelude::*;
 use futures::lock::Mutex;
+use hermes_any_counterparty::contexts::any_counterparty::AnyCounterparty;
 use hermes_async_runtime_components::subscription::impls::empty::EmptySubscription;
 use hermes_async_runtime_components::subscription::traits::subscription::Subscription;
-use hermes_cli_components::any_client::contexts::any_counterparty::AnyCounterparty;
 use hermes_cosmos_chain_components::components::client::*;
 use hermes_cosmos_chain_components::components::cosmos_to_cosmos::CosmosToCosmosComponents;
 use hermes_cosmos_chain_components::components::delegate::DelegateCosmosChainComponents;
