+++ crates/cosmos/cosmos-wasm-relayer/Cargo.toml
@@ -33,7 +33,7 @@ hermes-cosmos-chain-components          = { workspace = true }
 hermes-cosmos-test-components           = { workspace = true }
 hermes-cosmos-integration-tests         = { workspace = true }
 hermes-cosmos-relayer                   = { workspace = true }
-hermes-cli-components                   = { workspace = true }
+hermes-any-counterparty                 = { workspace = true }
 hermes-wasm-client-components           = { workspace = true }
 hermes-wasm-test-components             = { workspace = true }
